requirements.txt
async-timeout==3.0.1
discord==1.7.3
python-dotenv==0.20.0
dnspython==1.16.0
ffmpeg==1.4
typing-extensions==3.10.0.2
lyrics-extractor==3.0.1
youtube-dl==2021.12.17
PyNaCl==1.5.0
pipenv==2020.11.15